Split documentation into modules & add more documentation
